+++ jannovar-htsjdk/src/main/java/de/charite/compbio/jannovar/htsjdk/VariantContextAnnotator.java
@@ -286,7 +286,8 @@ public final class VariantContextAnnotator {
 
 		// If a VC builder is used before the attributes can be unmodifiable.
 		Map<String, Object> attributes = new HashMap<>(vc.getAttributes());
-		attributes.put("ANN", Joiner.on(',').join(annotations));
+		if (!annotations.isEmpty())
+			attributes.put("ANN", Joiner.on(',').join(annotations));
 		vc.getCommonInfo().setAttributes(attributes);
 
 		return vc;